Fixing issues and working on health bar

+++ Assets/Scripts/Interfaces/IAttack.cs
@@ -9,7 +9,7 @@ namespace Scripts.Interfaces
     {
         void Attack(bool attack);
         void Target(GameObject enemy);
-        int Damage();
+        float Damage();
     }
 }
